smart_anti_theft_tracker/requirements.txt
flask==2.3.3
cryptography==41.0.7
pyjwt==2.8.0
flask-sqlalchemy==3.0.5
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask-cors==4.0.1
flask-bcrypt==1.0.1
numpy==1.26.4  # Add for prediction math